+++ NEWS.md
@@ -1,6 +1,12 @@
 # mmrm 0.2.2.9000
 
 ### New Features
+- Add support for Kenward-Roger adjusted coefficients covariance matrix and
+  degrees of freedom in `mmrm` for spatial exponential covariance structure.
+
+# mmrm 0.2.2
+
+### New Features
 
 - Add support for Kenward-Roger adjusted coefficients covariance matrix and
   degrees of freedom in `mmrm` function call with argument `method`.